// wdip-fe/src/service/wdip-be/models/MotionsByParty.ts
// tslint:disable
/**
 * WDIP
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.0.1
 * Contact: markus@silberstein.nu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from '../runtime';
import {
    MotionsByPartyResults,
    MotionsByPartyResultsFromJSON,
    MotionsByPartyResultsToJSON,
} from './';

/**
 * 
 * @export
 * @interface MotionsByParty
 */
export interface MotionsByParty {
    /**
     * 
     * @type {string}
     * @memberof MotionsByParty
     */
    fromDate: string;
    /**
     * 
     * @type {string}
     * @memberof MotionsByParty
     */
    toDate: string;
    /**
     * 
     * @type {Array<MotionsByPartyResults>}
     * @memberof MotionsByParty
     */
    results: Array<MotionsByPartyResults>;
}

export function MotionsByPartyFromJSON(json: any): MotionsByParty {
    return {
        'fromDate': json['fromDate'],
        'toDate': json['toDate'],
        'results': (json['results'] as Array<any>).map(MotionsByPartyResultsFromJSON),
    };
}

export function MotionsByPartyToJSON(value?: MotionsByParty): any {
    if (value === undefined) {
        return undefined;
    }
    return {
        'fromDate': value.fromDate,
        'toDate': value.toDate,
        'results': (value.results as Array<any>).map(MotionsByPartyResultsToJSON),
    };
}